fix(deps)!: require transformers >= 5.3.0 for the models extra
Close Dependabot alerts CVE-2026-4372 (HIGH, config-injection RCE) and
CVE-2026-1839 (MEDIUM, Trainer torch.load RCE). Both are fixed only on the
transformers 5.x line -- CVE-2026-4372 first in 5.3.0 -- with no 4.x backport,
so the only version-based remediation is crossing the v4 -> v5 boundary, not a
downward cap. (The default install was never exposed anyway: the vulnerable
kernel-dispatch path needs the optional `kernels` package, which the extra does
not pull, and CVE-2026-1839 lives in Trainer, which the library never uses.)

MERT (at its pinned revision) and laion-clap 1.1.5 both run on transformers 5.x:
each imports only symbols that survive the v4 -> v5 refactor (verified against
the v5 Hubert / RoBERTa source). Pin MERT's load dtype to float32 explicitly,
since v5 changed the default to "auto" (would otherwise follow the checkpoint's
declared precision) and the engine is float32 end to end.

Raise the models-extra floors v5 mandates: torch >= 2.4 (was 2.1),
torchaudio >= 2.4, torchvision >= 0.19, huggingface_hub >= 1.0 (v5 switched the
hub HTTP backend to httpx). Lockfile refreshed alongside: transformers 5.13.0,
huggingface-hub 1.22.0, plus incidental dev/transitive bumps (mypy 2.2,
pot 0.9.7, filelock, charset-normalizer, librt).

The core (default, torch-free install) is unchanged and stays torch-free. The
torch-free suite is green; the real-model path (MERT/CLAP on transformers
5.13.0) is validated by the manual `test-models` job, not by the default CI.

BREAKING CHANGE: the [models] extra now requires transformers >= 5.3.0 and
torch >= 2.4 (previously transformers 4.x / torch 2.1). Consumers installing
moodengine[models] against a pinned transformers 4.x or torch < 2.4 must
upgrade those. The default install is unaffected.

+++ src/moodengine/embeddings/mert.py
@@ -71,8 +71,14 @@ class MERTEmbedder(Embedder):
             self.device,
         )
         try:
+            # transformers 5.x defaults the load dtype to "auto" (whatever precision
+            # the checkpoint declares); pin float32 explicitly so the engine's
+            # float32-end-to-end contract holds regardless of the model's config.
             self.model = AutoModel.from_pretrained(
-                config.mert_model_name, revision=revision, trust_remote_code=True
+                config.mert_model_name,
+                revision=revision,
+                trust_remote_code=True,
+                dtype=torch.float32,
             )
             self.feature_extractor = AutoFeatureExtractor.from_pretrained(
                 config.mert_model_name, revision=revision, trust_remote_code=True
